+++ src/sumologic-span-processor/override-xhr-span-name.ts
@@ -1,12 +1,12 @@
 import { Span } from '@opentelemetry/sdk-trace-base';
-import { isXhrInstrumentationSpan } from './utils';
+import { isXhrFetchSpan } from './utils';
 
 export const onStart = (span: Span): void => {
   const currentName = span.name;
 
   const updatedName = `HTTP ${currentName.toUpperCase()}`;
 
-  if (isXhrInstrumentationSpan(span)) {
+  if (isXhrFetchSpan(span)) {
     span.updateName(updatedName);
   }
 };